requirements.txt
gdown
transformers
pymongo
python-dotenv
fastapi
uvicorn==0.34.0
redis
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.7